Results table improvements
Moved results under command and parameter to address massive amounts of white space. Responses can be expanded and collapsed so that the user can easily scroll through all of the commands without having to scroll forever. Replaced unordered list format to remove extra spacing. Fixed size of table cells so that there isn't a large amount of space between a key and a value